Moved DuplicateAccount comp and small refactoring
